+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Bugfixes:
   - Updated `pouchdb-all-dbs` fixes [#2874](https://github.com/pubkey/rxdb/issues/2874)  Thanks [@openscript](https://github.com/openscript)
 
 Other:
+  - Determinstic handling of revision keys during data migration
   - Added more information to `RxError` when data migration fails
 
 ### 9.14.0 (14 February 2021)